added entity types and server side check if item exists in DB before proceed

+++ server/resolvers/resolvers.ts
@@ -1,4 +1,8 @@
-import { IResolvers, AuthenticationError } from "apollo-server-express";
+import {
+  IResolvers,
+  AuthenticationError,
+  UserInputError,
+} from "apollo-server-express";
 import BooksDB from "../models/book";
 import AuthorsDB from "../models/author";
 import UsersDB from "../models/user";
@@ -7,6 +11,7 @@ import {
   encryptPassword,
   comparePassword,
 } from "../config/authUtils";
+import { ErrorMessage } from "../types/types";
 
 type ObjType = Record<string, any>;
 
@@ -22,7 +27,7 @@ const resolvers: IResolvers = {
       if (context.loggedIn) {
         return context.user;
       } else {
-        throw new AuthenticationError("Please Login Again!");
+        throw new AuthenticationError(ErrorMessage.LOGIN_REQUIRED);
       }
     },
     users: async () => await UsersDB.find(),
@@ -38,25 +43,40 @@ const resolvers: IResolvers = {
 
   Mutation: {
     addAuthor: async (parent: ObjType, args: ObjType) => {
-      const author = new AuthorsDB({
+      const objExistingAuthor = await AuthorsDB.findOne({ name: args.name });
+      if (objExistingAuthor) {
+        throw new UserInputError(ErrorMessage.AUTHOR_EXISTS);
+      }
+
+      const objNewAuthor = new AuthorsDB({
         name: args.name,
         age: args.age,
       });
 
-      return await author.save();
+      return await objNewAuthor.save();
     },
 
     addBook: async (parent: ObjType, args: ObjType) => {
-      const book = new BooksDB({
+      const objExistingBook = await BooksDB.findOne({ name: args.name });
+      if (objExistingBook) {
+        throw new UserInputError(ErrorMessage.BOOK_EXISTS);
+      }
+
+      const objNewBook = new BooksDB({
         name: args.name,
         genre: args.genre,
         authorId: args.authorId,
       });
 
-      return await book.save();
+      return await objNewBook.save();
     },
 
     removeAuthor: async (parent: ObjType, args: ObjType) => {
+      const objExistingAuthor = await AuthorsDB.findById(args.id);
+      if (!objExistingAuthor) {
+        throw new UserInputError(ErrorMessage.AUTHOR_NOT_FOUND);
+      }
+
       // 1. remove all author's books (if any)
       await BooksDB.deleteMany({ authorId: args.id });
       // 2. remove Author:
@@ -64,17 +84,32 @@ const resolvers: IResolvers = {
     },
 
     removeBook: async (parent: ObjType, args: ObjType) => {
+      const objExistingBook = await BooksDB.findById(args.id);
+      if (!objExistingBook) {
+        throw new UserInputError(ErrorMessage.BOOK_NOT_FOUND);
+      }
+
       return await BooksDB.findByIdAndDelete(args.id);
     },
 
     editBook: async (parent: ObjType, args: ObjType) => {
       const { id, name, genre, authorId } = args;
+      const objExistingBook = await BooksDB.findById(id);
+      if (!objExistingBook) {
+        throw new UserInputError(ErrorMessage.BOOK_NOT_FOUND);
+      }
+
       await BooksDB.findByIdAndUpdate(id, { name, genre, authorId });
       return await BooksDB.findById(id);
     },
 
     editAuthor: async (parent: ObjType, args: ObjType) => {
       const { id, name, age } = args;
+      const objExistingAuthor = await AuthorsDB.findById(id);
+      if (!objExistingAuthor) {
+        throw new UserInputError(ErrorMessage.AUTHOR_NOT_FOUND);
+      }
+
       await AuthorsDB.findByIdAndUpdate(id, { name, age });
       return await AuthorsDB.findById(id);
     },
@@ -86,7 +121,7 @@ const resolvers: IResolvers = {
       });
 
       const objUser = await UsersDB.findOne({ username: args.username });
-      if (objUser) throw new AuthenticationError("User already exists.");
+      if (objUser) throw new AuthenticationError(ErrorMessage.USER_EXISTS);
 
       try {
         const objRegisteredUser = await objNewUser.save();
@@ -109,13 +144,13 @@ const resolvers: IResolvers = {
         const objFoundUser = { ...objUser?.toObject(), token: strToken };
         return objFoundUser;
       } else {
-        throw new AuthenticationError("wrong password.");
+        throw new AuthenticationError(ErrorMessage.WRONG_PASSWORD);
       }
     },
 
     removeUser: async (parent: ObjType, args: ObjType) => {
       const objUser = await UsersDB.findOne({ username: args.username });
-      if (!objUser) throw new AuthenticationError("User does not exist.");
+      if (!objUser) throw new AuthenticationError(ErrorMessage.USER_NOT_FOUND);
 
       try {
         const objDeletedUser = await objUser.deleteOne();
@@ -128,7 +163,7 @@ const resolvers: IResolvers = {
     editUser: async (parent: ObjType, args: ObjType) => {
       const { id, username, password } = args;
       const objUser = await UsersDB.findById(id);
-      if (!objUser) throw new AuthenticationError("User does not exist.");
+      if (!objUser) throw new AuthenticationError(ErrorMessage.USER_NOT_FOUND);
 
       try {
         const strNewPassword = await encryptPassword(password);
